backend/requirements.txt
fastapi==0.115.11
httpx==0.28.1
jose==1.0.0
numpy==2.2.4
pandas==2.2.3
pydantic==2.10.6
pydantic_settings==2.8.1
python-dotenv==1.0.1
python_bcrypt==0.3.2
python_jose==3.4.0
Requests==2.32.3
scikit_learn==1.6.1
SQLAlchemy==2.0.39
